Initialise check-in singletons under mutexes
This makes initialisation of the check-in scheduler (and the check-in
transmitter, for good measure) thread-safe, by having it take place
inside a mutex.

This fixes an issue where sending check-ins from different threads
simultaneously would cause several check-in schedulers to be
instantiated. On shutdown, only one of those check-in schedulers
would have its `#stop` method called, meaning check-in events
scheduled in the other schedulers would not be sent to AppSignal
when `Appsignal.stop` is called.