Set formatter on all broadcasted loggers
When a formatter is set for the broadcast logger, also set that same
formatter for all the loggers that it broadcasts to.

Before this change, the already-formatted message would be passed on
to the broadcasted loggers, which would format it once again.

This issue was masked by the tests matching the resulting log line
using `a_string_starting_with`, which I introduced in #1351 to smooth
out differences between our custom instrumentation of tagged logging,
since removed, and that of `ActiveSupport::TaggedLogging`. Ensure all
matching is done against specific strings.

Fixes issue #1360, in which the twice-applied formatting causes
additional newlines to appear in the message emitted by the
broadcasted loggers.