Instrument `ownership` gem
Add an instrumentation for the `ownership` gem. When a transaction
is created, if an owner has been set in the gem, use the name of
the owner as the owner tag for the transaction. Do the same when an
owner is set during an active transaction.

If more than one owner is set during a transaction, the owner tag
will be that of the last owner that was set.

When the `ownership_set_namespace` config option is enabled, set the
namespace for the transaction to the owner as well.