Do not report `SignalException` from Rake tasks
Same as #1267 for our `at_exit` error reporter, do not report
SignalException errors.

The `SignalException` can be a normal signal that leads to a shutdown.
We do not need to report those normal events.

If people do want to report this error they can add their own `at_exit`
handler, but let's not report it by default.