Deduplicate error wrappers and error causes
We got a report that an Action View error was reported twice. Once the
`ActionView::Template::Error` wrapper and once the error that actually
occurred in the view (the error cause, e.g. `NoMethodError`).

This was caused by receiving the error cause from the Rails error
reporter and the error wrapper in our error middleware. This was
detailed in issue https://github.com/appsignal/support/issues/345 and
Rails issue https://github.com/rails/rails/issues/53938.
(This only happened in production due to a Rails behavior quirk between
development and production, addressed in the issue I linked above.)

This change will make sure we do not report both errors, but only report
the `ActionView::Template::Error`.

Closes https://github.com/appsignal/support/issues/345