Add helper to create custom markers
Add a helper class that calls our Public Endpoint API to create custom
markers.

This helper is added to give people a method to report custom markers
from the Ruby gem. We have chosen to do this through the public endpoint
and not the Push API, because it's better fits the data flow we want.

It does not support reporting deploy markers, because we only want to
report those through the `revision` config option.

This request is authenticated with the Push API key, app name and app
environment as query parameters, which is supported by the Public
Endpoint API, similar to how we report check ins.
The 'normal' API requires a Personal Access Token, which gives a lot of
access to the reported data. This method uses a write only
authentication method.

Related work:

- Public Endpoint PR:
  https://github.com/appsignal/appsignal-processor-rs/pull/1679
- Docs PR:
  https://github.com/appsignal/appsignal-docs/pull/2419

Closes #1375