Improve _load_config! helper
Do not accept any empty or blank app environment strings.
It also casts symbols to strings.
This prevents any errors or weird behavior due to unexpected input.

Accept a block so that we can use this for the Capistrano config update.

Also document why we don't pass `context.env` to the
`load_dsl_config_file` helper, because I forgot again.