Refactor Capistrano specs
Standardize the Capistrano 2 and 3 specs a bit with each other.

I found that we don't assertion if we actually make marker requests to
the API. This change will make sure we do.

Move the `stub_marker_request` helper to the shared helpers, so we
have one helper for this, instead of one per spec file.