Remove Sidekiq ActiveJob case for job arguments
This case statement is never checked. The ActiveJob instrumentation is
wrapped around the Sidekiq middleware. This sets the job arguments
before Sidekiq can.

It's set with `add_params_if_nil` helper. Then when the Sidekiq
middleware wants to set parameters, they're already set, and the check
is skipped.

Remove it to avoid confusion about it.
Sidekiq 8 renamed this class to `Sidekiq::ActiveJob::Wrapper`, but we
don't need to check it.