Log AppSignal stopping as info
AppSignal being told to stop is an uncommon event that has
significant impact in the behaviour of the application past that
point, and as such I think `info` level logging is warranted.

Specifically, this seeks to make it easier to debug a common-ish
issue where customers interpret the `monitor_and_stop` method as
meaning "stop monitoring this transaction" and not "stop monitoring
AppSignal entirely".