Add config_error method to fetch error on config
To help out with #1389, add a way to get the error that occurred on
loading the config file available. This way people can check in the app
if an error occurred and reraise it or exit the app if necessary.

We don't want to add the behavior of raising the config error on start
by default, because the gem should never raise an error.

I decide not to put it behind a config option, because the only way to
do that is with an environment variable. This also gives more
flexibility for anyone who want to add any kind of custom behavior based
on the presence of an error.