Add Logger `<<` method
Adding this method improves our compatibility with Ruby's Logger class,
allowing it to be used in scenarios we previously didn't support.

One such scenario is with Rack's CommonLogger which uses this method to
write to the logger.

The method returns the character count, which is the same as the `<<`
method on the Ruby class.

Normally no formatting is applied. I've chosen to ignore that for now,
as we require a log severity, so that becomes 'info'.
To apply no further formatting, one should not set a formatter on our
logger.