Add check_if_started! method (#1395)
This method will provide a convenience method to raise an error if
AppSignal failed to start.

Some people prefer this behavior and this allows them to trigger that
behavior.

As previously mentioned, I don't want to add this to an environment
variable because the start up and config logic is complex enough. Last
thing we need is another scenario to consider.
This method call makes it very explicit what happens, where and by whom.