Only enable at_exit hook on containers
The reason we call `Appsignal.stop` is to flush data to our extension
before the app exits and hopefully before the container does.
It gives our extension and agent a bit more time to send the data before
they are shut down along with the container.

For VMs (non-container hosts or bare metal), this additional
`Appsignal.stop` call adds a half second delay to the shut down process
(see previous commit message), so it's not great to have it always add
that delay, even when there's no data to flush.

Part of #1387