Make enable_at_exit_hook more configurable
Instead of always calling it, give more options to control when it calls
`Appsignal.stop`.

At first this was a boolean, but then it would always call it or not
call it, even if there is an error. These new possible values gives us
the ability to set an 'on error' default that keeps the original
behavior.