Remove redundant cron check-in pairs
Before submitting a batch of check-in events, check for any pairs of
events that are redundant with other pairs of check-in events. This
would happen, for example, when running a `cron` block for a given
identifier many times in quick succession.

A redundant pair of cron check-in events is a pair of start and
finish cron check-in events for a given identifier, given that another
pair of start and finish check-in events exists for that identifier.

Unfortunately the same kind of check is not possible for calls to
`.cron` that do not use a block, which only send a finish check-in
event. This is because, when analyzing the list of check-in events,
it is not possible to know whether the finish event is meant to close
a check-in occurrence by itself (in which case it would be redundant
with other un-paired finish check-in events in the same timeframe) or
whether it is meant to close a start cron check-in event that was sent
in a previous batch.