Report the Active Job queue time (#1412)
When we instrument a job, record the queue time: the time from when it
was inserted into the queue until when it was picked up for execution.

This is the only way I found to report this metric, because Active Job
doesn't have a stats module like Sidekiq does.

The value itself needs to be reported as milliseconds, which is why it
performs a `* 1_000` calculation, because our metrics require durations
(time) to be sent in milliseconds.

I've updated the test implementation `enqueued_at` format to match the
actual format used by Active Job 8.

[Internal Slack thread](https://appsignal.slack.com/archives/C7XHXQ7N0/p1746715203918389)