Clear completed transactions if still in Thread
When a request changes the Ruby Fiber during a request, like when used
with Dry::Effect as reported in #1403, make sure the Ruby doesn't try to
reuse the completed transaction. This would make it report no data from
subsequent requests.

The completed transaction could stay in the
`Thread.current[:appsignal_transaction]` value store when this happens.
Check if the transaction is actually closed when we find an existing
transaction in memory.