Broadcast log messages regardless of level
When broadcasting log messages, do so regardless of the log level
of the AppSignal broadcaster. This allows loggers with less strict
levels to log messages according to their own level thresholds.

To avoid executing the block passed as a message several times, which
would run into the same issues as `ActiveSupport::BroadcastLogger`,
wrap the given message block with an object that can be cast into a
proc and be called more than once, but that only call the underlying
block once, with the same return value each time. This also ensures
that we don't call the underlying block if the message is not logged
by any of the broadcasted loggers.

When the logger is silenced, however, do not broadcast any logs over
the silenced log threshold. This makes sense because silencing is
about the application code that performs the logging, not about its
destination.

Move the `Exception` formatting to the `#add` method so that all
error levels, not just `#error`, can format exceptions.

Fixes #1422.