Add type signatures to the Ruby gem
Use the Sord gem to the project to generate the signatures based on the
YARD documentation.

Update the YARD documentation to documented previously undocumented
parts.

Mark a lot of private components with `@!visibility private` so it's not
generating signatures for it. Our customers don't need to see those
signatures as it's private APIs.

Marking this with `@api private` is not the same as `@!visibility
private`.