Update auth check return value
Instead of returning `nil` for an error, return the string `error`. This
will help with defining the return value in the YARD documentation and
makes it more clear from the return value what the result of the request
was.
I haven't found a way to define a tuple in YARD where one value is
either String or Nil.