Merge Hash-like values in sample data
Fix the issue reported in #1431 where first adding a Hash value and then
adding a ActiveSupport::HashWithIndifferentAccess value logs a warning.

Look if the values are not just the same class, but also if both values
are subclasses of the Hash, which makes them mergable.

I also added a test case for Sinatra's IndifferentHash, because it's a
very similar scenario.

Fix #1431