Only instrument Rails if AppSignal started
Issue #1442 can be reproduced without AppSignal starting or being
active in the current environment, because the Railtie will add
Rails instrumentation even if AppSignal is not active.

This should not happen -- no instrumentation should be added to the
customer's application if AppSignal is not active.

Unfortunately, it is not possible to modify the middleware chain at
`after_initialize`, so this only fixes the issue for `on_load`.