Fix streaming responses with Rack middleware
The `Rack::Events` middleware is not compatible with streaming bodies.
It wraps the body in a `Rack::Events::EventedBodyProxy`, which unlike
a `Rack::BodyProxy`, always responds to `#each`, even if the body it
is proxying does not respond to `#each`.

According to the Rack spec, when a body responds to both `#each` and
`#call`, the former is preferred, ultimately causing a `NoMethodError`
on the original body.

Deprecate the use of `Rack::Events` for our event-based Rack
instrumentation. Implement a modified `Appsignal::Rack::Events` which
wraps the body in a `Rack::BodyProxy`, ensuring it does not respond to
`#each` when the body it's proxying does not respond to `#each`.

Since `Appsignal::Rack::Events`' modifications break the `on_send`
event, we do not recommend its direct usage. Instead, we provide a
convenience `Appsignal::Rack::EventMiddleware` wrapper, which uses
our modified events middleware with our event handler.

A deprecation warning is emitted when `Appsignal::Rack::EventMiddleware`
is used with `Rack::Events`.