Monkey patch `NullHandle` for Rails 8.1
Monkey patch the `ActiveSupport::Notifications` `NullHandle` so that we
get events reported again.

The `NullHandle` is used when no subscribers are registered. That means
we get no events reported when there are no subscribers.
We're not a subscriber because of a performance problem in PR #150 we
decided instead to monkey patch it.

See issue #1471

## To do

- This requires unit tests.
- This requires testing in a real world Rails 8.1 app.