Truncate internal log lines (#1469)
* Truncate internal log lines

Truncate logs sent to our internal integrations logger.

We had a report that for a `ActionController::BadRequest` error the
appsignal.log file filled up with the uploaded file, as it's part of the
error message.
This will limit the log line length to 2000 characters. We can
increase this if needed.

Closes https://github.com/appsignal/support/issues/364

* Fix typo in changeset

Change truncate to truncated in the changeset.

[skip ci]

Co-authored-by: Martin Lipskis <lipskis@users.noreply.github.com>

---------

Co-authored-by: Martin Lipskis <lipskis@users.noreply.github.com>