Disable Ruby 3.3.4 from build matrix
CodeOwnership uses `oxidize-rb/actions/fetch-ci-data@v1` github action
to fetch a list of ruby versions to build native extension for. During
the realase of 2.0.0 it only returned ["3.2", "3.4"]. The gem fails to
load because there's no compiled native extension available for Ruby
3.3.

Missing builds for Ruby 3.3 seems to be a known issue in the
`oxidize-rb/actions` repo. See
https://github.com/oxidize-rb/actions/issues/62

Remove Ruby 3.3.4 from the build matrix.