Patch `build_handle` to intercept null handle
In Rails 8.1, `ActiveSupport::Notifications::Fanout` will return
the `NullHandle` module, instead of a `Handle` instance, when no
listeners are registered for an event. This breaks our event
instrumentation, which relies on monkey-patching the `Handle` class.

Fix this by intercepting the `build_handle` method in
`ActiveSupport::Notifications::Fanout`, and returning an instance
of a class that implements the same interface as `Handle` and
instruments the event.