Reduce duplication in ActiveSupport integrations
All the integrations in `ActiveSupportNotificationsIntegration` do
essentially the same thing, monkey-patching different bits of
`ActiveSupport::Notifications` to instrument the event that flows
through them, mostly differing on the specific module or class that
they instrument and the signature it expects.

Reduce duplication by implementing helper methods for starting an
event, given the event name, and finishing an event, given the
event name and payload.